+++ game/ato/flightstate/startup.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
 
-import logging
 from datetime import datetime, timedelta
 from typing import TYPE_CHECKING
 
 from .atdeparture import AtDeparture
 from .taxi import Taxi
 from ..starttype import StartType
-
-from game.settings.settings import FastForwardStopCondition
 
 if TYPE_CHECKING:
     from game.ato.flight import Flight
@@ -35,19 +32,6 @@ class StartUp(AtDeparture):
     @property
     def spawn_type(self) -> StartType:
         return StartType.COLD
-
-    def should_halt_sim(self) -> bool:
-        if (
-            self.flight.client_count > 0
-            and self.settings.fast_forward_stop_condition
-            == FastForwardStopCondition.PLAYER_STARTUP
-        ):
-            logging.info(
-                f"Interrupting simulation because {self.flight} has players and has "
-                "reached startup time"
-            )
-            return True
-        return False
 
     @property
     def description(self) -> str: