+++ game/ato/flightstate/taxi.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
 
-import logging
 from datetime import datetime, timedelta
 from typing import TYPE_CHECKING
 
 from .atdeparture import AtDeparture
 from .takeoff import Takeoff
 from ..starttype import StartType
-
-from game.settings.settings import FastForwardStopCondition
 
 if TYPE_CHECKING:
     from game.ato.flight import Flight
@@ -35,30 +32,6 @@ class Taxi(AtDeparture):
     @property
     def spawn_type(self) -> StartType:
         return StartType.WARM
-
-    def should_halt_sim(self) -> bool:
-        if (
-            self.flight.client_count > 0
-            and self.settings.fast_forward_stop_condition
-            == FastForwardStopCondition.PLAYER_TAXI
-        ):
-            logging.info(
-                f"Interrupting simulation because {self.flight} has players and has "
-                "reached taxi time"
-            )
-            return True
-        if (
-            self.settings.fast_forward_stop_condition
-            in {
-                FastForwardStopCondition.PLAYER_STARTUP,
-            }
-            and self.flight.client_count > 0
-        ):
-            logging.info(
-                f"Interrupting simulation because {self.flight} has players and is already taxiing "
-            )
-            return True
-        return False
 
     @property
     def description(self) -> str: