+++ game/sim/missionsimulation.py
@@ -2,14 +2,12 @@ from __future__ import annotations
 
 import json
 import time
-from datetime import timedelta
 from pathlib import Path
 from typing import Optional, TYPE_CHECKING
 
 from game.debriefing import Debriefing
 from game.missionarchive import archive_mission
 from game.missiongenerator import MissionGenerator
-from game.settings.settings import FastForwardStopCondition, CombatResolutionMethod
 from game.unitmap import UnitMap
 from .aircraftsimulation import AircraftSimulation
 from .missionresultsprocessor import MissionResultsProcessor
@@ -18,14 +16,6 @@ from ..profiling import logged_duration
 if TYPE_CHECKING:
     from game import Game
     from .gameupdateevents import GameUpdateEvents
-
-
-TICK = timedelta(seconds=1)
-
-
-class SimulationAlreadyCompletedError(RuntimeError):
-    def __init__(self) -> None:
-        super().__init__("Simulation already completed")
 
 
 class MissionSimulation:
@@ -43,42 +33,6 @@ class MissionSimulation:
     def begin_simulation(self) -> None:
         self.time = self.game.conditions.start_time
         self.aircraft_simulation.begin_simulation()
-
-    def tick(
-        self,
-        events: GameUpdateEvents,
-        combat_resolution_method: CombatResolutionMethod,
-        force_continue: bool,
-    ) -> GameUpdateEvents:
-        self.time += TICK
-        if self.completed:
-            raise RuntimeError("Simulation already completed")
-        if (
-            self.game.settings.fast_forward_stop_condition
-            == FastForwardStopCondition.DISABLED
-        ):
-            events.complete_simulation()
-            return events
-
-        # Stop fast forward if there are no clients and the settings require a player to reach a certain state.
-        if (
-            not self.game.ato_has_clients()
-            and self.game.settings.fast_forward_stop_condition
-            in {
-                FastForwardStopCondition.PLAYER_TAKEOFF,
-                FastForwardStopCondition.PLAYER_TAXI,
-                FastForwardStopCondition.PLAYER_STARTUP,
-                FastForwardStopCondition.PLAYER_AT_IP,
-            }
-        ):
-            events.complete_simulation()
-            return events
-
-        self.aircraft_simulation.on_game_tick(
-            events, self.time, TICK, combat_resolution_method, force_continue
-        )
-        self.completed = events.simulation_complete
-        return events
 
     def generate_miz(self, output: Path) -> None:
         with logged_duration("Mission generation"):